create single module for generating game module
